+++ test/web/mastodon_api/views/status_view_test.exs
@@ -36,7 +36,10 @@ defmodule Pleroma.Web.MastodonAPI.StatusViewTest do
     activity = Repo.get(Activity, activity.id)
     status = StatusView.render("show.json", activity: activity)
 
-    assert status[:pleroma][:emoji_reactions] == [["☕", 2], ["🍵", 1]]
+    assert status[:pleroma][:emoji_reactions] == [
+             %{emoji: "☕", count: 2},
+             %{emoji: "🍵", count: 1}
+           ]
   end
 
   test "loads and returns the direct conversation id when given the `with_direct_conversation_id` option" do